requirements.txt
# Temporary fork of llama-cpp-python to fix the llama.cpp download
llama-cpp-python

# OpenaI for Chat Completion API
openai

# Database
faiss-cpu

# Vector database
numpy
sentence-transformers

# Environment variables
python-dotenv

# Rust crate dependencies
nostr-sdk
nostr-protocol
